Make use of conftest.py
