+++ tests/test_users.py
@@ -4,21 +4,6 @@ from jose import jwt
 from app import schemas
 
 from app.config import settings
-from .database import client, session
-
-# def test_root(client):
-#     res = client.get("/")
-#     print(res.json().get("message"))
-#     assert res.json().get("message") == "Hello World"
-
-@pytest.fixture
-def test_user(client):
-    user_data = {"email": "hello123@gmail.com", "password": "password123"}
-    res = client.post("/users/", json=user_data)
-    assert res.status_code == 201
-    new_user = res.json()
-    new_user["password"] = user_data["password"]
-    return new_user
 
 
 def test_create_user(client):
